Crate rtfm_core

Core abstractions of the Real Time For the Masses (RTFM) framework

You can write generic libraries by using the Resource trait in this crate. If you want to write application code then you'll need an implementation of the RTFM framework for a particular architecture. Currently there are implementations for these two architectures:

ARM Cortex-M

MSP430

Structs

Threshold
Preemption threshold token

Traits

Resource
A resource, a mechanism to share data between tasks