Crate rtfm_core

Core abstractions of the Real Time For the Masses (RTFM) framework

You can write generic libraries using the Mutex trait in this crate. If you want to write application code then you'll need an implementation of the RTFM framework for a particular architecture. Currently, there are implementations for these architectures and OSes:

ARM Cortex-M

Structs

Exclusive
Newtype over &'a mut T that implements the Mutex trait

Traits

Mutex
Memory safe access to shared resources